+++ MediaPlayerWithTest.Controller/src/PlayListController.cs
@@ -13,16 +13,28 @@ namespace MediaPlayerWithTest.Controller.src
 
         public void AddNewFile(int playListId, int fileId, int userId)
         {
+            if (playListId <= 0 || fileId <= 0 || userId <= 0)
+            {
+                return;
+            }
             _playListService.AddNewFile(playListId, fileId, userId);
         }
 
         public void EmptyList(int playListId, int userId)
         {
+            if (playListId <= 0 || userId <= 0)
+            {
+                return;
+            }
             _playListService.EmptyList(playListId, userId);
         }
 
         public void RemoveFile(int playListId, int fileId, int userId)
         {
+            if (playListId <= 0 || fileId <= 0 || userId <= 0)
+            {
+                return;
+            }
             _playListService.RemoveFile(playListId, fileId, userId);
         }
     }
